set default cap hooks in a pre-deploy task
Issue: sidekiq default hooks are never added for cap3

Cause:
The cap3 sidekiq hooks are added based on a cap setting, the final value of which won't be available until "before deploy." Currently the check is done when the task file is loaded.

Solution:
Check and possibly add the hooks in a pre-deploy task rather than at the top level of the task file.