Use ActiveSupport.on_load hooks to extend rails components
Checking if the constant is define and including the extensions directly
in the base class will for the class to be loaded. This may cause these
components to be loaded before the time and be missconfigured if the
sidekiq railtie is loaded before the components railtie.

It is unlikely to happen but is better be safe.