Added method to Util module to clean up process records that don't have a matching heartbeat record, indicating that they're no longer alive.
Added call to this new util method in the Schedule::Poller.poll_interval
method because the number of live processes is used as a multiplier for
the default wait interval.  Since the value for poll_interval is
memoized this call to 'cleanup_dead_process_records' should only be
called once at startup.