concurrency value should also accept string as value
If we pass concurrency value as string it sidekiq fails to start.

We are sidekiq pro user and we use cloudformation templates to generate our UAT environments on Aws Opsworks. Now the problem is we  pass concurrency value as integer but the json that is generated by cloudformation for opsworks treat it as string and the yml that is generated for the sidekiq is like

```
---
:index: 0
:concurrency: '12'
:verbose: false
:timeout: 600
:queues:
- default
- queueone
- queuetwo
- queuethree
:logfile: log/sidekiq.log
```

because concurrency value is string it fails to start.