Don't shadow job method in Processor
The Processor class has a `#job` method which was being shadowed in
`#process` by a local variable. The `#process` method is complex, and I
recently moved some code around in a way that accidentally resulted in
the instance method being used instead of the local variable.

By using a different name for the variable, making that same mistake
again will result in a `NameError` instead of doing the wrong thing.