Change "really needs" to "must" in redis_connection.rb
This is just a semantic change, but if the program is going to crash, I think "must have" is more prescriptive than "really needs to have". 

In my mind, "really needs to have" implies that the program will continue run but the user should do something differently in the future.