client: bypass multi/exec when pushing single message
TLDR: this patch bypasses multi/exec when
pushing a single message in the sidekiq client,
which allows the client to be used with twemproxy.
this is extremely useful when running unicorn.

when running on unicorn, a common issue is
managing the number of connections to other
systems, e.g. redis. because each unicorn worker
runs in its own process, each process gets its
own connection, so in-process connection
pooling is not possible.

a solution to this conundrum is to use an
external pooling proxy such as twemproxy.

twemproxy brings with it some limitations,
however. some commands are not supported,
including multi/exec. so in order to be able
to use the sidekiq client with twemproxy, we
must not use multi/exec.

it turns out that when pushing a single message,
multi/exec is not needed, because in that case
we will always run only a single redis command.

atomic_push will either perform a zadd or an
sadd and an lpush. because the sadd is
idempotent, it should be safe.