Prefix every line in thread dump with TID
The previous behaviour relied on the order of lines in the produced
thread dump to match backtrace lines to threads. If you, for some
reason, can't rely on the order of log lines it becomes really hard to
make sense of the log.

Example: you transport your Sidekiq logs via UDP and once they're stored the
order is based on which UDP datagrams arrived first, not which line was
produced first.

To make debugging easier in cases like this, this change
1.) adds a the TID as a prefix to every printed line in the thread dump
2.) uses the separate log line for each line in the backtrace so they're
all prefixed with the local timestamps