Add unsafe-inline to allow bucket graph to display
The current CSP policy for Sidekiq Web appears to block the inline
script that renders the bucket charts for limits.

This change updates the script CSP rules to match the style CSP rules to
allow the chart to successfully render. In our application this change
made the bucket charts successfully render.

cc: #3913, #4030