Follow generator configuration to determine the resulting test file
Verifying that `RSpec` is defined assumes that the library is loaded on
the same environment that the generator is running.

It is often common to not have the rspec gem required on development,
which is the default environment to run the generator from.

Following the configured generator removes the assumption and works as
expected without breaking expectations.