Introduce the ability to pause queue from the UI (#4374)
* Introduce the ability to pause queue from the UI

Considering sidekiq web UI allow to perform complex tasks with ease.
Having the ability to pause a queue during an emergency (or operational intense)
scenarios would come in quite handy.

This PR introduces a new button to under the "Actions" column on `sidekiq/queues` pages,
right next to the `Delete` button. Depending on the state of the queue, it will according
show the `Pause` or `Unpause` text, with appropriate form element that `POST`s to the same
endpoint. The endpoint logic is updated to handle the new cases. Added some relevant unit tests.

Adding mock `Sidekiq::Queue#pause!` and `Sidekiq::Queue#unpause!` functions to
the API, which I believe will be overridden by sidekiq pro.

* Remove confirmation text and switch ordering of new copy in en.yml

* Only allow pause/unpause operations from sidekiq pro

This means the button from UI is hidden and api route has validation
in place such that it will only perform the respective pause/unpause
operation on proper payload, in sidekiq pro.

Its using `Sidekiq::Pro` constant as a check to detect whether or
not pro is present.

* Initialize Queue only once in api router action