Do not connect to redis at exit if not needed
A fix for https://github.com/mperham/sidekiq/issues/4498 introduced
a call-back that is executed at ruby VM exit and tries to connect to redis.

This call-back is also executed in CI runs and in all other cases where sidekiq is loaded but not used with real redis.

As a work-around, keep the at_exit hook but communicate to redis only if there is something to send, i.e. the number of processed or failed jobs is greater than zero