Add Queue#delete_jobs_from
This makes it a bit easier to cleanup and repair a queue from a bad job. Which
tends to be the most common case for using the Sidekiq::Queue#each method in my
experience. This is crucial when you're in a hurry and said bad job is
hammering a datastore to the point where it's impacting your whole application.

Before this addition, you had to do the following:

    queue.each { |job| job.delete if job.klass == "BadWorker" }

While definitely not bad, this is a bit of a contortion. My goal for this new
API was to be shorter to type than the original, have no block or conditional
and be easier to remember. Hence the following:

    queue.delete_jobs_from "BadWorker"

This also avoids having to know about the Sidekiq::Job API.