Fix `Uncaught exception: can't dump IO`
An error occurs when the underlying Redis client is configured with a
logger, due to the limitations of `Marshal` when dealing with `IO`
objects. This change excludes the `:logger` option in a similar fashion
to the way `:ssl_params` is currently excluded.

This error can be reproduced with the following:

    Sidekiq.configure_client do |config|
      config.redis = { logger: Logger.new(STDOUT) }
    end

The complete exception message and stacktrace is as follows:

    Uncaught exception: can't dump IO
    	/usr/local/bundle/gems/sidekiq-6.2.2/lib/sidekiq/redis_connection.rb:103:in `dump'
    	/usr/local/bundle/gems/sidekiq-6.2.2/lib/sidekiq/redis_connection.rb:103:in `log_info'
    	/usr/local/bundle/gems/sidekiq-6.2.2/lib/sidekiq/redis_connection.rb:32:in `create'
    	/usr/local/bundle/gems/sidekiq-6.2.2/lib/sidekiq.rb:137:in `redis='