Allow extensions to properly process empty hashes
In 0a4de94d76a07d9a61f0f55d87de79ff8bfe2f41 some existing
functionality was broken in the case of the sole argument being passed
to the exetension being an empty hash.

The **kwargs parameter interprets the hash as belonging to it, rather
than to args. The result of this is that when the `perform` method is
called, the args is now `[]` whereas before that commit, it was `[{}]`.

This commit restores the functionality by removing **kwargs and adding
code that replicates the functionality of **kwargs complete with a
special case for an empty hash being passed.