Reconnect on known errors after failover when pushing jobs to Redis (#5159)
In a Redis cluster setup, failovers will happen. In these
cases a `Redis::CommandError` can be raised for different
reasons, for example when the server becomes a replica, when
there is a "Not enough replicas" error from the primary, or
when a blocking command is force-unblocked.

These errors can occur when pushing a job to Redis, so it needs
to reconnect to the current master node and retry. Otherwise,
these jobs are lost.

The retry logic is similar to the implementation for `Sidekiq.redis`.