Extend perform_bulk to accept custom options hash similar to push_bulk
Some sidekiq plugins use `push_bulk` in sending batch jobs, along with some metadata to be passed
along the job that can later be read from the Sidekiq server middleware. However, such a thing is
not possible when with `perform_bulk`. It'd be really nice to support this where we can pass in the custom
options/attributes and get the benefit of batched redis calls too with `batch_size`

Ex:

```ruby
 Sidekiq::Client.push_bulk('class' => YourWorker, 'args' => args, 'tenant_id' => 3)
```

We can do the same with perform_bulk now

```ruby
  set = MySetJob.set('queue' => 'bar')
  set.perform_bulk(args, custom_opts: { 'tenant_id' => 3})
```