Restore compatibility with Rack 2
Response headers were previously downcased in order to support Rack 3.
However, this led to an issue where the content-security-policy header
was not found by the Rails 7 (Rack 2) ContentSecurityPolicy middleware
and the middleware would add its own (conflicting) header.

This commit restores compatibility with Rack 2 by making the casing of
response headers conditional on the Rack version. A few of these are
able to use the constants from Rack itself, and those not included in
Rack were added as constants under Sidekiq::Web.