Tighten CSP around `script-src` and `style-src`
Instead of allowing every src, require scripts from unknown sources to have have a nonce.
This makes it harder to exploit potential XSS vulnverabilities as the attacker needs to somehow know the nonce beforehand.

This also adds the nonce to all internal scripts. As per https://github.com/sidekiq/sidekiq/pull/3913#issuecomment-410318634 the assets may be hosted
outside of the main app. Adding the nonce continues to support this usecase.

A nonce is incompatible with `unsafe-inline` (which only style has). Adapt the chart to not assign inline styles directly.

Extensions that load external scripts/styles must either vendor or add the nonce to their tags.

Closes #6268