Make Sidekiq Web UI test less flaky (#6280)
The Sidekiq Web UI test attempts to test that the latency shown in the
queues page is 0.0. However, if there is some slowness in the test the
latency could tick up to a second, causing the test to fail.

Avoid this flakiness by stubbing `Time.now` so that we guarantee that
the latency will always remain at 0.