Do not fail on invalid locale in session hash
Prior to this commit, Sidekiq would treat the `locale` key in the
session as safe: If it's set, it's surely a supported locale. However,
in complex applications, the `locale` key in the session might be set by
another app to some value that is not what Sidekiq expects.

In this case, Sidekiq fails with a 404 error and no good explanation.

What this commit does is make sure that the value in `session[:locale]`
is actually a supported, available locale for Sidekiq. In case this is
an unwanted locale, a user can still use the locale selector to change
it back, but the Sidekiq admin doesn't suddenly become unreachable.

Another route this issue could be solved is by namespacing the `locale`
key in the session to something like `sidekiq_locale`, in order to be
safer from other parts of the app setting this value to something odd.
I'm opting for this approach though because I believe it to be a less
invasive change.