Demonstrate failing iterable batch job resumption
Currently the cursor is updated on each iteration before running the
iteration. This means that if the job is interrupted, the cursor will
point to the id of the record that was about to be processed.

When the job is resumed, if the enumerator returns a single item then
everything works fine because the cursor points to the only object we
need. However, for `ActiveRecordEnumerator#batches` and
`ActiveRecordEnumerator#relations` the cursor points at the last record
of the batch that hadn't been processed, so when the enumerator runs we
get a new batch starting from that record. That means we effectively
don't process all but the last record from the interrupted batch.

These new tests demonstrate the behaviour:

```
  1) Failure:
Sidekiq::Job::Iterable#test_0020_reschedules relations when sidekiq is
stopping [test/iterable/iterable_test.rb:251]:
--- expected
+++ actual
@@ -1 +1 @@
-[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+[1, 2, 3, 6, 7, 8, 9, 10]

  2) Failure:
Sidekiq::Job::Iterable#test_0019_reschedules batches when sidekiq is
stopping [test/iterable/iterable_test.rb:237]:
--- expected
+++ actual
@@ -1 +1 @@
-[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+[1, 2, 3, 6, 7, 8, 9, 10]

```