Move cursor to first id in batches and relations
Currently these ActiveRecordEnumerator methods set the cursor to the
last record in the batch. This means that if the job is stopped and
resumed we will skip all but the first record in the batch.

Instead, we should store the first record in the batch as the cursor so
the resumption will pick up where it left off.