Don't swallow NoMethodErrors in jobs setting current_attributes
If we have successfully set the current attributes, we don't retry if we
catch an error.

The `NoMethodError` that was being rescued from also applied to the job
that was being wrapped. This caused errors to be silently hidden and the
job retried without going into the retry queue, or alerting any bug
tracking tools that have been set up.