Web `/stats` endpoint does not return Valkey or Dragonfly Versions
We noticed that after migrating to Valkey and Sidekiq 8, the initial dashboard
page load would display the correct 8.0.1 version number for Valkey. But after
the auto-refresh of data, that number would switch back to 7.2.4.

This change introduces a JavaScript helper function that returns the correct
version number for the backing store. The order is the same as the behavior in
`#store_version` in helpers.rb.